return generateStaticParams
