--- src/app/racket/[racketId]/route.ts
@@ -1,1 +1,0 @@
-export { generateStaticParams } from './staticParams';
